feat: support clients without persistent cache
