+++ test/e2e/code/world.js
@@ -16,7 +16,7 @@ class PeerioAppWorld {
         this.parameters = parameters;
         this.filesToCleanup = [];
         this.contactsHelper = new ContactsHelper(this);
-        this.cacheStorage = {}; // for TestCacheEngine
+        this.cacheStorage = {}; // for MemoryCacheEngine
     }
 
     /**
